Unify AccessTracking unit tests and extract common to base class
Results of both System.getProperties() and System.getenv() are Maps so
there is a lot of intersection between tests. There are some corner
cases with regard to types (properties are Objects, env prohibits
anything but non-null String even in queries), but the vast majority of
test cases are common. This commit extracts all common tests to the
base class while rewriting them to follow the same pattern.